backend/requirements.txt
fastapi==0.115.0
pydantic==2.10.4
uvicorn[standard]==0.32.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pillow==10.1.0
reportlab==4.0.7
numpy==1.24.3
scipy==1.10.1
scikit-image==0.21.0
python-dotenv==1.0.0
sqlalchemy==2.0.23

# 水印相关库
invisible-watermark>=0.2.0
opencv-python==4.8.1.78
PyWavelets>=1.4.1
torch>=2.0.0  # 深度学习模型必需（可选）
torchvision>=0.15.0  # StegaStamp & Tree-Ring 必需（可选）

